backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
passlib==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
httpx==0.25.1
websockets==12.0
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.27.0
google-search-results==2.4.2
validators==0.22.0
pyaudio==0.2.13
pydub==0.25.1
librosa==0.10.0
soundfile==0.12.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
scipy==1.11.4
matplotlib==3.8.2
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
PyPDF2==3.0.1
python-docx==1.1.0
